slight refactor, expose sort reducer and setSort for use in parent store
